Add basic example
