+++ demo/lib/theme.dart
@@ -29,6 +29,12 @@ ThemeData get darkTheme {
         primary: Colors.purpleAccent,
         secondary: Colors.purpleAccent,
       ),
+      extensions: <ThemeExtension<dynamic>>[
+        const MyColors(
+          brandColor: Color(0xFF90CAF9),
+          danger: Color(0xFFEF9A9A),
+        ),
+      ],
     ),
   );
 }
@@ -41,6 +47,46 @@ ThemeData get lightTheme {
         primary: Colors.deepPurpleAccent,
         secondary: Colors.deepPurple,
       ),
+      extensions: <ThemeExtension<dynamic>>[
+        const MyColors(
+          brandColor: Color(0xFF1E88E5),
+          danger: Color(0xFFE53935),
+        ),
+      ],
     ),
   );
+}
+
+@immutable
+class MyColors extends ThemeExtension<MyColors> {
+  const MyColors({
+    required this.brandColor,
+    required this.danger,
+  });
+
+  final Color brandColor;
+  final Color danger;
+
+  @override
+  MyColors copyWith({Color? brandColor, Color? danger}) {
+    return MyColors(
+      brandColor: brandColor ?? this.brandColor,
+      danger: danger ?? this.danger,
+    );
+  }
+
+  @override
+  MyColors lerp(ThemeExtension<MyColors>? other, double t) {
+    if (other is! MyColors) {
+      return this;
+    }
+    return MyColors(
+      brandColor: Color.lerp(brandColor, other.brandColor, t)!,
+      danger: Color.lerp(danger, other.danger, t)!,
+    );
+  }
+
+  // Optional
+  @override
+  String toString() => 'MyColors(brandColor: $brandColor, danger: $danger)';
 }
