+++ demo/lib/views/basic_example.dart
@@ -1,3 +1,4 @@
+import 'package:demo/theme.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:mix/mix.dart';
@@ -23,38 +24,4 @@ class BasicExample extends HookWidget {
       child: const TextMix('Gradient Box'),
     );
   }
-}
-
-@immutable
-class MyColors extends ThemeExtension<MyColors> {
-  const MyColors({
-    required this.brandColor,
-    required this.danger,
-  });
-
-  final Color brandColor;
-  final Color danger;
-
-  @override
-  MyColors copyWith({Color? brandColor, Color? danger}) {
-    return MyColors(
-      brandColor: brandColor ?? this.brandColor,
-      danger: danger ?? this.danger,
-    );
-  }
-
-  @override
-  MyColors lerp(ThemeExtension<MyColors>? other, double t) {
-    if (other is! MyColors) {
-      return this;
-    }
-    return MyColors(
-      brandColor: Color.lerp(brandColor, other.brandColor, t)!,
-      danger: Color.lerp(danger, other.danger, t)!,
-    );
-  }
-
-  // Optional
-  @override
-  String toString() => 'MyColors(brandColor: $brandColor, danger: $danger)';
 }
